requirements.txt
google-cloud
google-api-python-client
google-auth-httplib2
oauth2client